Refactoring is performed, correcting scripts to capture.
