--- extract/ciarp/__init__.py
@@ -1,1 +1,0 @@
-"""CIARP extractor package."""
