--- extract/doaj/__init__.py
@@ -1,5 +1,0 @@
-"""DOAJ extractor package."""
-
-from .doaj_extractor import DoajExtractor
-
-__all__ = ["DoajExtractor"]
