+++ tests/etl/test_ciarp.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 import mongomock
 import pytest
 
-from extract.ciarp.ciarp_extractor import CiarpExtractor
+from dags.ciarp_capture import CiarpExtractor
 
 # Workaround for mongomock bug with pymongo 4.x bulk_write (seen in some bulk ops)
 # https://github.com/mongomock/mongomock/issues/812
@@ -16,6 +16,15 @@ if hasattr(mongomock.collection.BulkOperationBuilder, "add_update"):
         return original_add_update(self, filter, doc, upsert, multi, **kwargs)
 
     mongomock.collection.BulkOperationBuilder.add_update = patched_add_update
+
+if hasattr(mongomock.collection.BulkOperationBuilder, "add_replace"):
+    original_add_replace = mongomock.collection.BulkOperationBuilder.add_replace
+
+    def patched_add_replace(self, selector, doc, upsert=False, **kwargs):
+        kwargs.pop("sort", None)
+        return original_add_replace(self, selector, doc, upsert=upsert, **kwargs)
+
+    mongomock.collection.BulkOperationBuilder.add_replace = patched_add_replace
 
 
 @pytest.fixture
